apps/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.2
psycopg[binary]==3.2.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9
aiofiles==23.2.1
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
itsdangerous==2.2.0
argon2-cffi==23.1.0
python-magic==0.4.27
loguru==0.7.2
